requirements.txt
torch
transformers
diffusers
accelerate
tqdm
torchvision
Pillow
